requirements.txt
torch
numpy
matplotlib
torchvision
argparse
thop
pycocotools
Pillow
opencv-python
